Update after review
Review comments:
- https://github.com/ibm-granite/granite-io/pull/67#pullrequestreview-2683477132

Main points:
- Restructure of input and output process under io/granite_3_2
- Rename granite_3-2.py files to be differentiated
- Need to validate that input passed to output processor is Granite 3.2 valid
