+++ src/granite_io/io/granite_3_2/input_processors/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""
+The input processor module holds the abstraction and implementations for
+Granite 3.2 input processing.
+"""
+
+# Local
+import granite_io.io.granite_3_2.input_processors.granite_3_2_input_processor  # noqa: F401
