--- src/granite_io/io/output_processors/__init__.py
@@ -1,11 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-The output processors module holds the abstraction and implementations for model
-output processing.
-"""
-
-# Local
-from granite_io.io.output_processors.granite_3_2.granite_3_2 import (  # noqa: F401
-    Granite3Point2OutputProcessor,
-)
