--- tests/io/output_processors/granite_3_2/testdata/test_output.txt
@@ -1,27 +1,0 @@
-To efficiently find the fastest way for a seller to visit all cities in their region, consider implementing the following steps and methods: 
-
-1. **Data Collection**: The first step involves gathering and structuring data about all cities the seller needs to visit. This data should include each city's geographical coordinates (latitude and longitude) or precise addresses to calculate exact distances accurately.
-
-2. **Distance Matrix Creation**: Compute the pairwise distances between every combination of cities. The most common approach is using the Haversine formula for great-circle distances if dealing with spherical coordinates (latitude and longitude). Alternatively, you could use driving distance APIs like Google Maps API or OpenRouteService, which provide actual travel distances considering road networks.
-
-3. **Choose an Algorithm**: Depending on the number of cities and required precision:
-
-   - **Small Number of Cities (<20-30)**: Exact algorithms like the branch-and-bound method or dynamic programming can be used to find the optimal solution. `tsp_solver` library in Python, for instance, utilizes a branch-and-bound algorithm implemented by Lin-Kernighan.
-   
-   - **Larger Number of Cities (>30)**: Heuristic and metaheuristic methods often provide good approximations in reasonable time:
-
-     - **Nearest Neighbor**: A simple greedy approach where the salesman starts at a city and always moves to the nearest unvisited city. Though not guaranteed optimal, it’s fast and can serve as a baseline.
-     
-     - **Genetic Algorithms**: These mimic natural selection processes to evolve better solutions over generations. They're especially useful for complex, large-scale problems but require tuning parameters like mutation rate and population size. 
-     
-     - **Ant Colony Optimization (ACO)**: An algorithm inspired by ant behavior that gradually constructs a solution through iterative improvements guided by pheromone trails. ACO libraries are available in Python (pyacs) and other languages.
-     
-   - **Simulation-based Methods**: If specific time windows for visits exist, constrained optimization techniques or simulation methods (like vehicle routing with time windows) could be employed.
-
-4. **Implementation and Testing**: Implement the chosen method using appropriate libraries/software (Python's `tsp`, Concorde TSP solver, or custom code). Validate your results against simpler cases or known small problem instances to ensure correctness.
-
-5. **Optimization & Fine-tuning**: Based on test results, refine parameters or tweak methods for better performance. Consider parallelization/distributed computing for large datasets to speed up calculations.
-
-6. **Result Interpretation**: Once you have the route ( Sequence of cities), calculate total distance or estimated time based on average travel speeds between cities to quantify the 'fastest' path.
-
-Without knowing the exact number of cities and data availability, this outline provides a flexible approach applicable across various scales of the problem. For very large datasets or real-time applications, cloud-based services or specialized TSP solvers might be beneficial.
