--- tests/io/output_processors/granite_3_2/testdata/test_output_with_citation_from_source.txt
@@ -1,6 +1,0 @@
-The team that has shown the most success in the last decade of hurling is Limerick. In the specified timeframe, they won the All-Ireland SHC three times (2020, 2021, and 2022), the Munster SHC four times (2019, 2020, 2021, and 2022), and the National Hurling League twice as well (in 2019 and 2020). This consistent achievement makes Limerick stand out among other hurling teams.
-
-<co>1</co>
-
-# Citations:
-<co>1</co> Document 0: "The 2018 season concluded with Limerick winning the 2018 All-Ireland SHC, the team's first since 1973, with a 3–16 to 2–18 point defeat of Galway in the final.The team built on this success, winning the NHL in 2019, 2020 and 2023, the Munster SHC in 2019, 2020, 2021 and 2022 and the All-Ireland SHC again in 2020, 2021 and 2022. Munster Senior Hurling Championship 2023, All Ireland Hurling Championship 2023 to be forever remembered the team to join the Cork hurling Champions of the 40s and the Kilkenny hurling Champions of the 2000s to complete 4 in a row."
