--- tests/io/output_processors/granite_3_2/testdata/test_output_with_citation_hallucinations.txt
@@ -1,9 +1,0 @@
-RAG stands for Retrieval-Augmented Generation. It's a method that enhances the abilities of generative artificial intelligence models by integrating them with information retrieval capabilities <co>1</co>. This technique, mentioned in Document 1, allows these AI models to access and utilize external data during their generation process, potentially leading to more accurate and contextually rich outputs.
-
-# Citations:
-<co>1</co> Document 1: "RAG, retrieval-augmented generation, is a technique that grants generative artificial intelligence models information retrieval capabilities."
-
-# Hallucinations:
-1. Risk low: RAG stands for Retrieval-Augmented Generation.
-2. Risk low: It's a method that enhances the abilities of generative artificial intelligence models by integrating them with information retrieval capabilities <co>1</co>.
-3. Risk low: This technique, mentioned in Document 1, allows these AI models to access and utilize external data during their generation process, potentially leading to more accurate and contextually rich outputs.
