--- tests/io/output_processors/granite_3_2/testdata/test_output_with_invalid_citation.txt
@@ -1,4 +1,0 @@
-To provide a contrast between extreme technological determinism and social constructivism with a more systemic understanding of how technology and society co-produce each other <c>1</c>  <c>2</c>. This introduction highlights the importance of co-production as a mode of knowledge formation that acknowledges the interplay between technology and society.
-
-# Citations:
-<c>1</c> Document 1: "Co-production is a form of knowledge production based on the dynamic interaction between technology and society; where technical experts and other groups come together, with their different ways of viewing and analyzing the world and, in the process, generate new knowledge and technologies.Andrew L. Christenson. The Co-production of Archaeological Knowledge: The Essential Relationship of Amateurs and Professionals in 20th Century American Archaeology. Complutum (Universidad Complutense de Madrid) 24(2): 63-72, 2013 It has a long history, particularly arising out of radical theories of knowledge in the 1970s and, beyond science and technology studies, is often applied to public services and administration and forms the basis of participatory development.\n\nDescription\nCo-production forms part of Mode 2,Michael Gibbons, Camille Limoges, Helga Nowotny, Simon Schwartzman, Peter Scott and Martin Trow The New Production of Knowledge: The dynamics of science and research in contemporary societies Sage. 1994Jasanoff, Sheila. States of Knowledge: The Co-Production of Science and the Social Order. Routledge. 2004. Harbers, Hans. Inside the Politics of Technology: Agency and Normativity in the Co-Production of Technology and Society. Amsterdam University Press. 2005.  a term used in the sociology of science to describe one of the modes\u2014or ways\u2014that knowledge is formed. In Mode 2, science and technology studies move from extreme technological determinism and social constructivism, to a more systemic understanding of how technology and society \u2018co-produce\u2019 each other. Co-production is functionally comparable to the concepts of causality loop, positive feedback, and co-evolution \u2013 all of which describe how two or more variables of a system affect and essentially create each other, albeit with respect to different variables operating at different scales. And as with these other concepts, if used too broadly/uncritically, co-production risks noetic flatness \u2013 if technology and society co-produce each other equally, the justification for maintaining the boundary between them dissolves (in which case actor-network theory may be invoked). Unless overlapping sets of boundary-work are employed, co-production may also fail to account for power differentials within each variable, (in this case, within technology and society).\n\nScience, technology and society\nFrom a more science, technology and society (STS) perspective, Sheila Jasanoff, has written that \"Co-production is shorthand for the proposition that the ways in which we know and represent the world (both nature and society) are inseparable from the ways in which we chose to live in.\" Co-production draws on constitutive (such as Actor\u2013network theory) and interactional work (such as the Edinburgh School) in STS. As a sensitizing concept, the idiom of co-production looks at four themes: \"the emergence and stabilization of new techno-scientific objects and framings, the resolution of scientific and technical controversies; the processes by which the products of techno-science are made intelligible and portable across boundaries; and the adjustment of science\u2019s cultural practices in response to the contexts in which science is done.\" Studies employing co-production often follow the following pathways: \"making identities, making institutions, making discourses, and making representations\"Jasanoff, Sheila. States of Knowledge: The Co-Production of Science and the Social Order . Routledge. 2004\n\n Co-production of climate services \nA disconnect exists between the climate information that is produced by science (in terms of weather forecasts and climate projections) and what is needed by users to make climate-resilient decisions. The mismatch usually relates to time scales, spatial scales, and metrics. Co-producing climate services, by bringing together producers and users of climate information for dialogue, can lead to the creation of new knowledge that is more appropriate for use in terms of being tailored and targeted to particular decisions.\n\nAs in other fields, co-production of climate services, can create challenges due to differences in the incentives, priorities and languages of the various parties (often grouped into \"producers\" of information and \"users\" of information). Although there are no recipes for how to co-produce climate services, there are a number of building blocks and principles. Building blocks for co-production of climate services are:\n\n identifying key actors and building partnerships\n building common ground\n co-exploring need\n co-developing solutions\n co-delivering solutions\n evaluation.\n\nPrinciples for co-production of climate services are:\n\n improve transparency of forecast accuracy and certainty\n tailor to context and decision\n deliver timely a sustainable services\n build trust\n embrace diversity and respect differences\n enhance inclusivity\n keep flexible\n support conscious facilitation\n communicate in accessible ways\n ensure value-add for all involved.\n\nReferences\n\nExternal links\nExamples of initiatives to co-produce climate services include:\n Weather and Climate Information Services for East Africa (WISER)\n Future Climate For Africa (FCFA) \n Enhancing National Climate Services initiative (ENACTS).\n\nCategory:Science studies\nCategory:Science and technology studies\n\nja:\u5354\u50cd\n\nWhy does the author introduce the concept of co-production as a part of Mode 2 in the sociology of science?"
